research/archive/legacy-flask-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
rdkit-pypi==2022.9.5
pickle-mixin==1.0.2
gunicorn==21.2.0
python-dotenv==1.0.0
